解決多數人無法註冊的問題 原因: 1. pvp points 是 float 型態 2. post 無法獲取資料，這個要在post前先用get拿資料

+++ DiscordBot/Repositories/KogDataModel.cs
@@ -5,7 +5,7 @@ public class PointType
     public int Rank { get; set; }
     public string Name { get; set; } = default!;
     public int TPoints { get; set; }
-    public int PvPpoints { get; set; }
+    public float PvPpoints { get; set; }
     public int Points { get; set; }
     public int Seasonpoints { get; set; }
     public int RewardIndex { get; set; }
